+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.4.3
 - Support for DZRP 1.6.0. CMD_CLOSE implemented.
+- startAutomatically with z80 unit tests improved.
 
 # 1.4.2
 - Support for DZRP 1.4.0: Changed numbering.